+++ app/src/main/java/com/kaizen/bangunpc/ui/screen/catalog/CatalogScreen.kt
@@ -15,7 +15,7 @@ import com.kaizen.bangunpc.ui.theme.AppTheme
 
 @Composable
 fun CatalogScreen(
-    navigateToDetailProduct: (Long) -> Unit = {}
+    navigateToDetailProduct: (Int) -> Unit = {}
 ) {
     Column {
         CustomTopBar(
